jar: Refactor manifest writing into a ManifestResource class
We place the manifest writing logic into a ManifestResource class
which allows adding the manifest to a Jar as a resource and get
proper output.
